Rename feature ui item
