+++ WordPress/src/main/java/org/wordpress/android/ui/debug/DebugSettingsItemViewHolder.kt
@@ -10,10 +10,10 @@ import android.widget.TextView
 import androidx.annotation.LayoutRes
 import androidx.recyclerview.widget.RecyclerView.ViewHolder
 import org.wordpress.android.R
-import org.wordpress.android.ui.debug.DebugSettingsViewModel.FeatureUiItem
-import org.wordpress.android.ui.debug.DebugSettingsViewModel.FeatureUiItem.Feature.State.DISABLED
-import org.wordpress.android.ui.debug.DebugSettingsViewModel.FeatureUiItem.Feature.State.ENABLED
-import org.wordpress.android.ui.debug.DebugSettingsViewModel.FeatureUiItem.Feature.State.UNKNOWN
+import org.wordpress.android.ui.debug.DebugSettingsViewModel.UiItem
+import org.wordpress.android.ui.debug.DebugSettingsViewModel.UiItem.Feature.State.DISABLED
+import org.wordpress.android.ui.debug.DebugSettingsViewModel.UiItem.Feature.State.ENABLED
+import org.wordpress.android.ui.debug.DebugSettingsViewModel.UiItem.Feature.State.UNKNOWN
 
 sealed class DebugSettingsItemViewHolder(
     parent: ViewGroup,
@@ -21,14 +21,14 @@ sealed class DebugSettingsItemViewHolder(
 ) : ViewHolder(LayoutInflater.from(parent.context).inflate(layout, parent, false)) {
     class HeaderViewHolder(parent: ViewGroup) : DebugSettingsItemViewHolder(parent, R.layout.manual_feature_header) {
         private val header = itemView.findViewById<TextView>(R.id.header)
-        fun bind(item: FeatureUiItem.Header) {
+        fun bind(item: UiItem.Header) {
             header.setText(item.header)
         }
     }
 
     class ButtonViewHolder(parent: ViewGroup) : DebugSettingsItemViewHolder(parent, R.layout.manual_feature_button) {
         private val button = itemView.findViewById<Button>(R.id.button)
-        fun bind(item: FeatureUiItem.Button) {
+        fun bind(item: UiItem.Button) {
             button.setText(item.text)
             button.setOnClickListener { item.clickAction() }
         }
@@ -38,7 +38,7 @@ sealed class DebugSettingsItemViewHolder(
         private val title = itemView.findViewById<TextView>(R.id.feature_title)
         private val enabled = itemView.findViewById<CheckBox>(R.id.feature_enabled)
         private val unknown = itemView.findViewById<ImageView>(R.id.unknown_icon)
-        fun bind(item: FeatureUiItem.Feature) {
+        fun bind(item: UiItem.Feature) {
             title.text = item.title
             enabled.visibility = View.GONE
             unknown.visibility = View.GONE
